requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
websockets
aiofiles
sqlalchemy
passlib[bcrypt]
python-jose
python-dotenv
requests